provided templates and components are now located in the CodeInc\UI\Library namespace
