+++ src/ComponentInterface.php
@@ -19,70 +19,20 @@
 // Project:  UI
 //
 declare(strict_types=1);
-namespace CodeInc\UI\Component\Library;
-use CodeInc\UI\Component\ComponentInterface;
-
+namespace CodeInc\UI;
 
 /**
- * Class StringComponent
+ * Interface ComponentInterface
  *
- * @package CodeInc\UI\Component\Library
+ * @package CodeInc\UI
  * @author Joan Fabrégat <joan@codeinc.fr>
  */
-class StringComponent implements ComponentInterface
+interface ComponentInterface
 {
     /**
-     * @var string
-     */
-    private $content;
-
-    /**
-     * Adds some content.
-     *
-     * @param string $content
-     */
-    public function add(string $content):void
-    {
-        $this->content .= $content;
-    }
-
-    /**
-     * Sets or replaces the content.
-     *
-     * @param string $content
-     */
-    public function set(string $content):void
-    {
-        $this->content = $content;
-    }
-
-    /**
-     * Verifies if the content is empty.
-     *
-     * @return bool
-     */
-    public function isEmpty():bool
-    {
-        return empty($this->content);
-    }
-
-    /**
-     * Returns the content's length.
-     *
-     * @return int
-     */
-    public function length():int
-    {
-        return strlen($this->content);
-    }
-
-    /**
-     * Returns the content as a string.
+     * Returns the content's generated output.
      *
      * @return string
      */
-    public function get():string
-    {
-        return $this->content;
-    }
+    public function get():string;
 }
