+++ src/Library/Templates/BlankHtmlTemplate.php
@@ -3,57 +3,65 @@
 // +---------------------------------------------------------------------+
 // | CODE INC. SOURCE CODE                                               |
 // +---------------------------------------------------------------------+
-// | Copyright (c) 2018 - Code Inc. SAS - All Rights Reserved.           |
+// | Copyright (c) 2017 - Code Inc. SAS - All Rights Reserved.           |
 // | Visit https://www.codeinc.fr for more information about licensing.  |
 // +---------------------------------------------------------------------+
 // | NOTICE:  All information contained herein is, and remains the       |
 // | property of Code Inc. SAS. The intellectual and technical concepts  |
 // | contained herein are proprietary to Code Inc. SAS are protected by  |
 // | trade secret or copyright law. Dissemination of this information or |
-// | reproduction of this material is strictly forbidden unless prior    |
+// | reproduction of this material  is strictly forbidden unless prior   |
 // | written permission is obtained from Code Inc. SAS.                  |
 // +---------------------------------------------------------------------+
 //
 // Author:   Joan Fabrégat <joan@codeinc.fr>
-// Date:     08/10/2018
+// Date:     20/02/2018
+// Time:     15:21
 // Project:  UI
 //
-declare(strict_types=1);
-namespace CodeInc\UI\Component\Library;
-use CodeInc\UI\Component\ComponentInterface;
-
+namespace CodeInc\UI\Library\Templates;
 
 /**
- * Class ImmutableStringComponent
+ * Class BlankHtmlTemplate
  *
- * @package CodeInc\UI\Component\Library
+ * @package CodeInc\UI\Library\Templates
  * @author Joan Fabrégat <joan@codeinc.fr>
  */
-class ImmutableStringComponent implements ComponentInterface
+class BlankHtmlTemplate extends AbstractHtmlTemplate
 {
     /**
-     * @var string
-     */
-    private $string;
-
-    /**
-     * ImmutableStringComponent constructor.
-     *
-     * @param string $string
-     */
-    public function __construct(string $string)
-    {
-        $this->string = $string;
-    }
-
-
-    /**
-     * Returns the content's generated output.
-     *
+     * @inheritdoc
      * @return string
      */
-    public function get():string
-    {
-        return $this->string;
-    }
+	protected function getTemplateHeader():string
+	{
+	    $lang = $this->getLanguage();
+	    ob_start();
+		?>
+		<!DOCTYPE html>
+		<html<?=$lang ? ' lang="'.htmlspecialchars($lang).'"' : ''?>>
+			<head>
+				<meta charset="<?=htmlspecialchars($this->getCharset())?>">
+				<title><?=htmlspecialchars($this->getTitle())?></title>
+				<?=$this->getHeaders()->get()?>
+			</head>
+
+			<body>
+		<?
+        return ob_get_clean();
+	}
+
+    /**
+     * @inheritdoc
+     * @return string
+     */
+	public function getTemplateFooter():string
+	{
+	    ob_start();
+		?>
+			</body>
+		</html>
+		<?
+        return ob_get_clean();
+	}
 }
